+++ src/cache/KeyValueHeap.ts
@@ -1,24 +1,30 @@
-import { KeyValueCacheMap } from "./KeyValueCacheMap.js"
 import { CacheItemIndex, orderByExpiredTSScoreFunction, CachedValue } from "@/types/index.js"
-import { Integer } from "@/util/CommonTypes.js"
+import { Integer, NullableNumber } from "@/util/CommonTypes.js"
 import { IMapStorage } from "./IMapStorage.js"
 import MinHeap from "min-heap"
+import { Queue } from "@/util/Queue.js"
+import { IKeyValueCache } from "@/cache/IKeyValueCache.js"
+import { MapStorageImpl } from "@/cache/MapStorageImpl.js"
+import { isPositiveInteger } from "@/util/CommonConstrains.js"
+import { max } from "lodash"
 
 /**
- * A extant of MAP which use a min-heap to store a index of item's expireTS
- * to achieve faster expired item management.
+ * Use a min-heap to store a index of item's expireTS to achieve faster expired item management.
  * Support arbitrary ttl.
  *
  * !!! Important !!!
  * delete(key) won't remove the index as it require a O(N) operation to search for the index with the given key.
  * The index (of deleted item) only got remove when clearExpiredItems() / deleteFirstExpiredItem() / rebuildIndex() is called.
  */
-export class KeyValueCacheHeap<V> extends KeyValueCacheMap<V> {
+export class KeyValueCacheHeap<V> extends IKeyValueCache<V> {
+    private _store: IMapStorage<V>
+
     /**
      * The index for all cache item that will expire (has ttl & expireTS)
      * except the one that will expire first (_smallestExpiredTSItem)
      */
     private _itemExpiredTSIndexHeap: MinHeap<CacheItemIndex> = new MinHeap(orderByExpiredTSScoreFunction)
+
     /**
      * The cache item which will expire first.
      * For fast check if every items are not expired yet.
@@ -27,8 +33,52 @@ export class KeyValueCacheHeap<V> extends KeyValueCacheMap<V> {
      */
     private _smallestExpiredTSItem: CacheItemIndex | null = null
 
+    private _insertOrder: Queue<CacheItemIndex> = new Queue<CacheItemIndex>()
+
+    /**
+     * @param defaultTTL default time to live in seconds
+     * @param maxSize maximum number of items in the cache
+     * @param storage the storage object to use
+     */
     constructor(defaultTTL?: Integer, maxSize?: Integer, storage?: IMapStorage<V>) {
-        super(defaultTTL, maxSize, false, storage)
+        super(defaultTTL, maxSize)
+        if (storage) {
+            // use provided storage object
+            this._store = storage
+        } else {
+            // otherwise, use default storage which is a Map
+            this._store = new MapStorageImpl<V>()
+        }
+    }
+
+    protected getStoredItem(key: string): CachedValue<V> | undefined {
+        return this._store.get(key)
+    }
+
+    public hasExpired(key: string): boolean | undefined {
+        let cacheValue: CachedValue<V> | undefined = this.getStoredItem(key)
+        if (cacheValue) {
+            return CachedValue.hasExpired(cacheValue)
+        }
+        return undefined
+    }
+
+    /**
+     * See parent.
+     * Complexity: O(1)
+     */
+    get(key: string): V | undefined {
+        let cacheValue: CachedValue<V> | undefined = this.getStoredItem(key)
+        if (cacheValue) {
+            // if in cache, check if expired
+            if (CachedValue.hasExpired(cacheValue)) {
+                this.delete(key)
+                return undefined
+            } else {
+                return cacheValue.value
+            }
+        }
+        return undefined
     }
 
     /**
@@ -37,48 +87,92 @@ export class KeyValueCacheHeap<V> extends KeyValueCacheMap<V> {
      *      - O(1) to insert the value
      *      - O(log N) to insert the index
      *      - Overall: O(log N)
-     * @param key
-     * @param value
-     * @param ttl
      */
-    put(key: string, value: V, ttl?: number | undefined): void {
-        super.put(key, value, ttl)
-        let expireTS = this.getStoredItem(key)?.expireTS
+    put(key: string, value: V, ttl?: NullableNumber): void {
+        // validation
+        if (!value) {
+            throw new Error("Value cannot be a null")
+        }
 
-        if (expireTS != undefined) {
+        if (ttl && !isPositiveInteger(ttl)) {
+            throw new Error("Timeout is not a number, or less then or equal to 0")
+        } else if (ttl === null) {
+            ttl = this.defaultTTL
+        }
+
+        // create new item
+        let newValue: CachedValue<V> = new CachedValue(value, ttl)
+
+        // delete the old item with the same key (if any), ignore its expireTS
+        this._store.delete(key)
+        this._store.set(key, newValue)
+
+        // insert to index
+        let item = this.getStoredItem(key) as CachedValue<V>
+        let cacheItemIndex = new CacheItemIndex(key, item.insertTS, item.expireTS)
+
+        this._insertOrder.enqueue(cacheItemIndex)
+        if (item.expireTS != undefined) {
             // this item will expire
             if (this._smallestExpiredTSItem == null) {
                 // no current smallest, take the place
-                this._smallestExpiredTSItem = new CacheItemIndex(key, expireTS)
+                this._smallestExpiredTSItem = cacheItemIndex
             } else {
-                if (expireTS < this._smallestExpiredTSItem.expiredTS) {
+                if (item.expireTS < this._smallestExpiredTSItem.expiredTS) {
                     // this item should be the smallest
 
                     // put the current smallest back to the heap
                     this._itemExpiredTSIndexHeap.insert(this._smallestExpiredTSItem)
 
                     // take the place of the smallest
-                    this._smallestExpiredTSItem = new CacheItemIndex(key, expireTS)
+                    this._smallestExpiredTSItem = cacheItemIndex
                 } else {
                     // the smallest remains
                     // put the new item to the heap
-                    this._itemExpiredTSIndexHeap.insert(new CacheItemIndex(key, expireTS))
+                    this._itemExpiredTSIndexHeap.insert(cacheItemIndex)
                 }
             }
         }
+
+        if (this.isOverflow()) {
+            // remove overflow item
+            this.deleteFirstExpiredItem()
+
+            // check again if still overflow
+            if (this.isOverflow()) {
+                this.deleteFirstInsertedItem()
+            }
+        }
+
+        if (this.isIndexTooBig()) {
+            // if the index is too big, rebuild it
+            this.rebuildIndex()
+        }
     }
 
-    indexSize(): Integer {
-        return this._itemExpiredTSIndexHeap.size
+    isIndexTooBig(): boolean {
+        let indexSize = max([this._itemExpiredTSIndexHeap.size, this._insertOrder.size()]) as number
+        return indexSize > 2 * this.size()
+    }
+
+    /**
+     * Complexity: O(1)
+     */
+    delete(key: string): boolean {
+        return this._store.delete(key)
     }
 
     /**
      * Clear the cache values and the index
      */
     clear(): void {
-        super.clear()
+        this._store.clear()
         this._itemExpiredTSIndexHeap.clear()
         this._smallestExpiredTSItem = null
+    }
+
+    size(): Integer {
+        return this._store.size()
     }
 
     /**
@@ -111,6 +205,27 @@ export class KeyValueCacheHeap<V> extends KeyValueCacheMap<V> {
             }
             this._smallestExpiredTSItem = this._itemExpiredTSIndexHeap.removeHead()
         }
+        console.log("Deleted one item", this._smallestExpiredTSItem?.key)
+    }
+
+    deleteFirstInsertedItem(): void {
+        let deleteOneItem = false
+        while (!deleteOneItem) {
+            // no item is deleted as no item is expired
+            // delete the first inserted item
+            let key = this._insertOrder.dequeue()
+            if (key) {
+                let cacheValue = this.getStoredItem(key.key)
+                if (cacheValue && cacheValue.insertTS == key.insertTS) {
+                    // is the same item
+                    this.delete(key.key)
+                    break
+                }
+            } else {
+                // no item to delete
+                break
+            }
+        }
     }
 
     /**
@@ -118,33 +233,31 @@ export class KeyValueCacheHeap<V> extends KeyValueCacheMap<V> {
      */
     rebuildIndex(): void {
         let newHeap: MinHeap<CacheItemIndex> = new MinHeap(orderByExpiredTSScoreFunction)
+        let newInsertOrder: Queue<CacheItemIndex> = new Queue<CacheItemIndex>()
         let currentCacheItem: CachedValue<V> | undefined
 
-        // for each item in the old heap index,
+        // for each item in the old insert order
         // check if the item is still in the cache
-        // if yes, put it to the new heap
+        // if yes, put it to the new heap & insert order
 
-        // put the smallest expired item back to the heap
-        if (this._smallestExpiredTSItem != null) {
-            this._itemExpiredTSIndexHeap.insert(this._smallestExpiredTSItem)
-        }
-
-        while (this._itemExpiredTSIndexHeap.size > 0) {
-            let item = this._itemExpiredTSIndexHeap.removeHead()
-            currentCacheItem = this.getStoredItem(item.key)
-            if (currentCacheItem != undefined) {
-                if (CachedValue.hasExpired(currentCacheItem)) {
-                    // the item has expired, don't put it back
-                    this.delete(item.key)
-                } else if (item.expiredTS === currentCacheItem.expireTS) {
-                    // the index is still for the same item
-                    newHeap.insert(item)
+        while (this._insertOrder.size() != 0) {
+            let key = this._insertOrder.dequeue()
+            if (key) {
+                currentCacheItem = this.getStoredItem(key.key)
+                if (
+                    currentCacheItem &&
+                    currentCacheItem.insertTS == key.insertTS &&
+                    !CachedValue.hasExpired(currentCacheItem)
+                ) {
+                    newHeap.insert(key)
+                    newInsertOrder.enqueue(key)
                 }
             }
         }
 
         // replace the old heap with the new one
         this._itemExpiredTSIndexHeap = newHeap
+        this._insertOrder = newInsertOrder
         this._smallestExpiredTSItem = this._itemExpiredTSIndexHeap.removeHead()
     }
 }