+++ src/config/index.ts
@@ -1,2 +1,1 @@
 export * from "./CacheOption.js"
-export * from "./QueueConfig.js"