+++ src/types/CacheItemIndex.ts
@@ -1,15 +1,18 @@
-import { Timestamp } from "@/util/CommonTypes.js";
+import { Timestamp } from "@/util/CommonTypes.js"
 
 /**
  * Index to keep in the heap / queue.
  * Where expiredTS is the order key and key is used to id the cache item from the cache store.
  */
 export class CacheItemIndex {
-    private _expiredTS: Timestamp;
-    private _key: string;
+    private static oneHundredYears = 100 * 365 * 24 * 60 * 60 * 1000
+    private _expiredTS: Timestamp
+    private _insertTS: Timestamp
+    private _key: string
 
-    constructor(key: string, expiredTS: number){
-        this._expiredTS = expiredTS
+    constructor(key: string, insertTS: number, expiredTS?: number) {
+        this._insertTS = insertTS
+        this._expiredTS = expiredTS ?? this._insertTS + CacheItemIndex.oneHundredYears
         this._key = key
     }
 
@@ -20,33 +23,41 @@ export class CacheItemIndex {
      * @returns positive if this item expire later than another
      */
     compare(another: CacheItemIndex): number {
-        return this.expiredTS - another.expiredTS;
+        return this.expiredTS - another.expiredTS
     }
 
     /**
      * Getter expiredTS
      * @return {Timestamp}
      */
-	public get expiredTS(): Timestamp {
-		return this._expiredTS;
-	}
+    public get expiredTS(): Timestamp {
+        return this._expiredTS
+    }
+
+    /**
+     * Getter insertTS
+     * @return {Timestamp}
+     */
+    public get insertTS(): Timestamp {
+        return this._insertTS
+    }
 
     /**
      * Getter key
      * @return {string}
      */
-	public get key(): string {
-		return this._key;
-	}
+    public get key(): string {
+        return this._key
+    }
 }
 
 /**
- * The function to pass in to the constructor of the heap. 
+ * The function to pass in to the constructor of the heap.
  * So the the index can be order by the expireTS in ascending.
- * @param itemA 
- * @param itemB 
- * @returns 
+ * @param itemA
+ * @param itemB
+ * @returns
  */
-export function orderByExpiredTSScoreFunction(itemA: CacheItemIndex, itemB: CacheItemIndex){
-    return itemA.compare(itemB);
+export function orderByExpiredTSScoreFunction(itemA: CacheItemIndex, itemB: CacheItemIndex) {
+    return itemA.compare(itemB)
 }