+++ src/types/index.ts
@@ -1,4 +1,2 @@
-export * from "./CacheType.js"
 export * from "./CachedValue.js"
-export * from "./TimeoutMode.js"
 export * from "./CacheItemIndex.js"