+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Refactor the logic related the deleteFirst() in MapStorage
   - as the original method is not necessarily removing the first expired item
 - Better index size management
+- Remove unnecessary implementation of IKeyValueCache
 
 ## 1.2.0
 - Update to ES Module